patch 8.2.3132: compiler warns for size_t to colnr_T conversion.
Problem:    Compiler warns for size_t to colnr_T conversion. (Randall W.
            Morris)
Solution:   Add a type cast.

+++ src/drawscreen.c
@@ -2036,7 +2036,7 @@ win_update(win_T *wp)
 			{
 			    colnr_T t;
 
-			    pos.col = STRLEN(ml_get_buf(wp->w_buffer,
+			    pos.col = (int)STRLEN(ml_get_buf(wp->w_buffer,
 							     pos.lnum, FALSE));
 			    getvvcol(wp, &pos, NULL, NULL, &t);
 			    if (toc < t)
